requirements.txt
Django==1.9.2
django-crispy-forms==1.6.0
django-material==0.6.0
django-summernote==0.8.5
django-tinymce==2.3.0
django-wysiwyg==0.8.0
Pillow==3.3.1
pkg-resources==0.0.0
psycopg2==2.6.2
XlsxWriter==0.9.1
